chore: match 컴포넌트화
